Prevent "Unknown app_id" crashes
applicationSentListing is invoked when tab is changed. Sometimes there can be situation when the tab's app id is not in the hash table
It can occur when Safari was in background when proxy was launched or Safari goes into background and comes back again (SafariLauncher).
When such situation occur just ignore the update and don't crash the proxy